modify video clip utils

+++ src/dataset/utils/video.py
@@ -326,13 +326,15 @@ def read_video(
     aframes_list = [frame.to_ndarray() for frame in audio_frames]
 
     if vframes_list:
-        vframes = torch.as_tensor(np.stack(vframes_list))
+        # vframes = torch.as_tensor(np.stack(vframes_list))
+        vframes = torch.as_tensor(np.stack(vframes_list), dtype=torch.uint8)
     else:
         vframes = torch.empty((0, 1, 1, 3), dtype=torch.uint8)
 
     if aframes_list:
         aframes = np.concatenate(aframes_list, 1)
-        aframes = torch.as_tensor(aframes)
+        # aframes = torch.as_tensor(aframes)
+        aframes = torch.as_tensor(aframes, dtype=torch.float32)
         if pts_unit == "sec":
             start_pts = int(math.floor(start_pts * (1 / audio_timebase)))
             if end_pts != float("inf"):
